fix(todos): expose dashboard in normal host plugin entry
