+++ plugins/plugin-todos/src/index.ts
@@ -4,35 +4,6 @@
  * schema, and the `TodosView` dashboard view) and re-exports its types, service,
  * schema, and views. Hard-depends on `@elizaos/plugin-sql`.
  */
-import type { Plugin } from "@elizaos/core";
-import { todosRuntimePlugin } from "./plugin.js";
-
-export const todosPlugin: Plugin = {
-  ...todosRuntimePlugin,
-  views: [
-    {
-      id: "todos",
-      label: "Todos",
-      description: "Three-lane todo board: Today / Upcoming / Someday",
-      icon: "ListChecks",
-      path: "/todos",
-      responseContext: { primaryContext: "todos" },
-      modalities: ["gui"],
-      bundlePath: "dist/views/bundle.js",
-      // First-party instrumented view (data-agent-id controls): grant the
-      // agent-surface capability so the view broker admits agent-driven
-      // fills/clicks (#13452 manifest gate).
-      surface: { capabilities: ["agent-surface"] },
-      componentExport: "TodosView",
-      tags: ["todos", "tasks", "productivity"],
-      relatedActions: ["OWNER_TODOS"],
-      visibleInManager: true,
-      desktopTabEnabled: true,
-    },
-  ],
-};
-
-export default todosPlugin;
 
 export { todoAction } from "./actions/todo.js";
 export {
@@ -52,7 +23,11 @@ export {
   todosSchema,
   todosTable,
 } from "./db/schema.js";
-export { todosRuntimePlugin } from "./plugin.js";
+export {
+  todosPlugin,
+  todosPlugin as default,
+  todosRuntimePlugin,
+} from "./plugin.js";
 export { currentTodosProvider } from "./providers/current-todos.js";
 export {
   type CreateTodoInput,
